requirements.txt
PyQt5>=5.15.0
pyqtgraph>=0.13.0
numpy>=1.21.0
Pillow>=9.0.0
PyOpenGL>=3.1.0
